+++ typescript/sdk/src/token/Token.ts
@@ -77,6 +77,7 @@ import {
   RadixNativeTokenAdapter,
   RadixTokenAdapter,
 } from './adapters/RadixTokenAdapter.js';
+import { SealevelHypCrossCollateralAdapter } from './adapters/SealevelCrossCollateralAdapter.js';
 import {
   SealevelHypCollateralAdapter,
   SealevelHypNativeAdapter,
@@ -346,6 +347,18 @@ export class Token implements IToken {
         token: collateralAddressOrDenom,
         mailbox,
       });
+    } else if (standard === TokenStandard.SealevelHypCrossCollateral) {
+      assert(mailbox, `Mailbox required for Sealevel hyp tokens`);
+      assert(
+        collateralAddressOrDenom,
+        `collateralAddressOrDenom required for Sealevel hyp cross-collateral tokens`,
+      );
+
+      return new SealevelHypCrossCollateralAdapter(chainName, multiProvider, {
+        warpRouter: addressOrDenom,
+        token: collateralAddressOrDenom,
+        mailbox,
+      });
     } else if (standard === TokenStandard.SealevelHypSynthetic) {
       assert(mailbox, `Mailbox required for Sealevel hyp tokens`);
       assert(
@@ -553,7 +566,8 @@ export class Token implements IToken {
   isCrossCollateralToken(): boolean {
     return (
       this.standard === TokenStandard.EvmHypCrossCollateralRouter ||
-      this.standard === TokenStandard.TronHypCrossCollateralRouter
+      this.standard === TokenStandard.TronHypCrossCollateralRouter ||
+      this.standard === TokenStandard.SealevelHypCrossCollateral
     );
   }